fix: keep ExpireSnapshots retention defaults at int64 on 32-bit builds
Properties.GetInt parses to platform-width int, so math.MaxInt is only
MaxInt32 on 32-bit targets. The ExpireSnapshots retention defaults
(MaxSnapshotAgeMsDefault, MaxRefAgeMsDefault) are math.MaxInt sentinel
values that get truncated to ~25 days in milliseconds before the int64()
widening, causing snapshots older than 25 days to be expired even when
no retention was configured.

Add Properties.GetInt64 (parse with bit size 64, return the full int64)
and retype the two age defaults to int64(math.MaxInt64).

Closes #1278.

+++ table/transaction.go
@@ -312,11 +312,11 @@ func (t *Transaction) ExpireSnapshots(opts ...ExpireSnapshotsOpt) error {
 	// Read table-level retention properties as the last-resort defaults,
 	// mirroring the Java implementation. When neither the ref nor the
 	// caller provides a value, fall back to the table property; when the
-	// table property is also absent use the constant default (math.MaxInt,
+	// table property is also absent use the constant default (math.MaxInt64,
 	// meaning "keep everything").
-	propMaxRefAgeMs := int64(t.meta.props.GetInt(MaxRefAgeMsKey, MaxRefAgeMsDefault))
+	propMaxRefAgeMs := t.meta.props.GetInt64(MaxRefAgeMsKey, MaxRefAgeMsDefault)
 	propMinSnapshotsToKeep := t.meta.props.GetInt(MinSnapshotsToKeepKey, MinSnapshotsToKeepDefault)
-	propMaxSnapshotAgeMs := int64(t.meta.props.GetInt(MaxSnapshotAgeMsKey, MaxSnapshotAgeMsDefault))
+	propMaxSnapshotAgeMs := t.meta.props.GetInt64(MaxSnapshotAgeMsKey, MaxSnapshotAgeMsDefault)
 
 	for refName, ref := range t.meta.refs {
 		// Assert that this ref's snapshot ID hasn't changed concurrently.
